refactor(helm): extract newInstallAction, move test-only hooks, fix name collision
